requirements.txt
fastapi==0.116.1
uvicorn
python-multipart
pillow==10.4.0
numpy==2.1.1
open_clip_torch==3.0.0
torch==2.5.1
torchaudio==2.5.1
torchinfo==1.8.0
torchsummary==1.5.1
torchvision==0.20.1
PyYAML